requirements.txt
flask
pymongo[srv]
python-dotenv
jurigged
openpyxl
flask-cors
